+++ packages/react-zeugma/src/entities/zeugma/model/context.ts
@@ -1,5 +1,0 @@
-import { createContext } from 'react'
-import { ZeugmaStateValue, ZeugmaActionsValue } from './types'
-
-export const ZeugmaStateContext = createContext<ZeugmaStateValue | undefined>(undefined)
-export const ZeugmaActionsContext = createContext<ZeugmaActionsValue | undefined>(undefined)