+++ packages/react-zeugma/src/entities/zeugma/model/index.ts
@@ -1,6 +1,4 @@
-export * from './types'
-export * from './context'
 export * from './hooks'
-export * from './portal-context'
 export * from './useZeugma'
 export * from './useZeugmaContext'
+export * from './useZeugmaDnd'