+++ packages/react-zeugma/src/shared/lib/index.ts
@@ -1,3 +1,4 @@
 export * from './tree'
 export * from './drag-session'
 export * from './json'
+export * from './context'